move encrypt/decrypt code to NoiseSession
