+++ crypto.go
@@ -1,23 +1,13 @@
 package noise
 
-import (
-	"errors"
-	"github.com/libp2p/go-libp2p-noise/core"
-)
+import "errors"
 
 func (s *secureSession) Encrypt(plaintext []byte) (ciphertext []byte, err error) {
 	if !s.xx_complete && !s.ik_complete {
 		return nil, errors.New("decrypt err: haven't completed handshake")
 	}
 
-	if s.initiator {
-		cs := s.ns.CS1()
-		_, ciphertext = core.EncryptWithAd(cs, nil, plaintext)
-	} else {
-		cs := s.ns.CS2()
-		_, ciphertext = core.EncryptWithAd(cs, nil, plaintext)
-	}
-
+	ciphertext = s.ns.Encrypt(plaintext)
 	return ciphertext, nil
 }
 
@@ -26,18 +16,5 @@ func (s *secureSession) Decrypt(ciphertext []byte) (plaintext []byte, err error)
 		return nil, errors.New("decrypt err: haven't completed handshake")
 	}
 
-	var ok bool
-	if s.initiator {
-		cs := s.ns.CS2()
-		_, plaintext, ok = core.DecryptWithAd(cs, nil, ciphertext)
-	} else {
-		cs := s.ns.CS1()
-		_, plaintext, ok = core.DecryptWithAd(cs, nil, ciphertext)
-	}
-
-	if !ok {
-		return nil, errors.New("decrypt err: could not decrypt")
-	}
-
-	return plaintext, nil
+	return s.ns.Decrypt(ciphertext)
 }
